게시물 수정, 삭제

+++ src/main/java/com/study/market/board/service/BoardServiceImpl.java
@@ -1,6 +1,7 @@
 package com.study.market.board.service;
 
-import com.study.market.board.domain.Board;
+import com.study.market.auth.SecurityUtil;
+import com.study.market.board.domain.entity.Board;
 import com.study.market.board.domain.ResponseRetrieveBoardDto;
 import com.study.market.board.repository.BoardRepository;
 import com.study.market.item.repository.ItemRepository;
@@ -12,7 +13,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -20,7 +20,7 @@ import java.util.Map;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class BoardServiceImpl implements BoardService{
+public class BoardServiceImpl implements BoardService {
 
     private final BoardRepository boardRepository;
     private final MemberRepository memberRepository;
@@ -30,7 +30,7 @@ public class BoardServiceImpl implements BoardService{
     @Transactional
     public void saveBoard(String title, String content, String writer) {
         Member member = memberRepository.findByEmail(writer)
-                .orElseThrow(()->new IllegalArgumentException("해당 이메일이 존재하지 않습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("해당 이메일이 존재하지 않습니다."));
 
         Board board = Board.builder()
                 .title(title)
@@ -47,7 +47,7 @@ public class BoardServiceImpl implements BoardService{
     @Transactional
     public void modifyBoard(Long boardId, String title, String content) {
         Board board = boardRepository.findById(boardId)
-                .orElseThrow(()->new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
 
         board.changeTitle(title);
         board.changeContent(content);
@@ -59,7 +59,18 @@ public class BoardServiceImpl implements BoardService{
     @Transactional
     public void deleteBoard(Long boardId) {
         Board board = boardRepository.findById(boardId)
-                .orElseThrow(()->new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+
+        /*
+        Long memberId = SecurityUtil.getCurrentMemberId();
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
+
+        if (board.getMember().getId() != member.getId()) {
+            throw new IllegalArgumentException("해당 게시글을 삭제할 권한이 없습니다.");
+        }
+        //게시글 작성자만 삭제 가능 -> 작성자에게만 삭제 버튼이 보여서 이 부분은 필요 없음
+         */
 
         boardRepository.delete(board);
     }
@@ -71,7 +82,7 @@ public class BoardServiceImpl implements BoardService{
         //boardList에서 title만 추출한 리스트인 boardName 반환
         Map<Long, String> boardNameMap = new HashMap();
 
-        for(Board board : boardList){
+        for (Board board : boardList) {
             boardNameMap.put(board.getId(), board.getTitle());
         }
 
@@ -82,15 +93,33 @@ public class BoardServiceImpl implements BoardService{
     @Transactional
     public ResponseRetrieveBoardDto retrieveBoard(Long boardId) {
         Board board = boardRepository.findById(boardId)
-                .orElseThrow(()->new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
 
-        board.addViewCount();
+        Long memberId = SecurityUtil.getCurrentMemberId();
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
+
+        if (board.getMember().getId() != member.getId()) {
+            board.addViewCount(); //본인이 아닌 사람의 게시물 볼 때만 조회수 증가
+
+            return ResponseRetrieveBoardDto.builder()
+                    .title(board.getTitle())
+                    .content(board.getContent())
+                    .writer(board.getMember().getEmail())
+                    .viewCount(board.getViewCount())
+                    .isWriter("false")
+                    .build();
+        }
 
         return ResponseRetrieveBoardDto.builder()
                 .title(board.getTitle())
                 .content(board.getContent())
                 .writer(board.getMember().getEmail())
                 .viewCount(board.getViewCount())
+                .isWriter("true")
                 .build();
     }
+
+
+
 }
